attendance_portal/requirements.txt
alabaster==0.7.12
Babel==2.7.0
certifi==2019.9.11
chardet==3.0.4
commonmark==0.9.1
Django==2.2.6
django-crispy-forms==1.8.0
django-widget-tweaks==1.4.5
djangorestframework==3.10.3
docutils==0.15.2
idna==2.8
imagesize==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
packaging==19.2
Pygments==2.4.2
pyparsing==2.4.2
pytz==2019.3
recommonmark==0.6.0
requests==2.22.0
six==1.12.0
snowballstemmer==2.0.0
Sphinx==2.2.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
urllib3==1.25.6
